+++ saleor/plugins/base_plugin.py
@@ -157,6 +157,12 @@ class BasePlugin:
         ["User", str, str, Optional[str], None], None
     ]
 
+    # Trigger when account change email is requested.
+    #
+    # Overwrite this method if you need to trigger specific logic after an account
+    # change email is requested.
+    account_change_email_requested: Callable[["User", str, str, str, str, None], None]
+
     # Trigger when account delete is requested.
     #
     # Overwrite this method if you need to trigger specific logic after an account