+++ saleor/plugins/webhook/plugin.py
@@ -151,26 +151,30 @@ class WebhookPlugin(BasePlugin):
             )
 
     def _trigger_account_event(
-        self, event_type, user, channel_slug, token, redirect_url
+        self, event_type, user, channel_slug, token, redirect_url, new_email=None
     ):
         if webhooks := get_webhooks_for_event(event_type):
-            payload = self._serialize_payload(
-                {
-                    "id": graphene.Node.to_global_id("User", user.id),
-                    "token": token,
-                    "redirect_url": redirect_url,
-                }
-            )
+            raw_payload = {
+                "id": graphene.Node.to_global_id("User", user.id),
+                "token": token,
+                "redirect_url": redirect_url,
+            }
+            data = {
+                "user": user,
+                "channel_slug": channel_slug,
+                "token": token,
+                "redirect_url": redirect_url,
+            }
+
+            if new_email:
+                raw_payload["new_email"] = new_email
+                data["new_email"] = new_email
+
             trigger_webhooks_async(
-                payload,
+                self._serialize_payload(raw_payload),
                 event_type,
                 webhooks,
-                {
-                    "user": user,
-                    "channel_slug": channel_slug,
-                    "token": token,
-                    "redirect_url": redirect_url,
-                },
+                data,
                 self.requestor,
             )
 
@@ -190,6 +194,26 @@ class WebhookPlugin(BasePlugin):
             channel_slug,
             token,
             redirect_url,
+        )
+
+    def account_change_email_requested(
+        self,
+        user: "User",
+        channel_slug: str,
+        token: str,
+        redirect_url: str,
+        new_email: str,
+        previous_value: None,
+    ) -> None:
+        if not self.active:
+            return previous_value
+        self._trigger_account_event(
+            WebhookEventAsyncType.ACCOUNT_CHANGE_EMAIL_REQUESTED,
+            user,
+            channel_slug,
+            token,
+            redirect_url,
+            new_email,
         )
 
     def account_delete_requested(