+++ saleor/plugins/webhook/tests/subscription_webhooks/subscription_queries.py
@@ -28,6 +28,35 @@ ACCOUNT_CONFIRMATION_REQUESTED = (
 )
 
 
+ACCOUNT_CHANGE_EMAIL_REQUESTED = (
+    fragments.CUSTOMER_DETAILS
+    + """
+    subscription{
+      event{
+        ...on AccountChangeEmailRequested{
+          user{
+            ...CustomerDetails
+          }
+          token
+          redirectUrl
+          channel{
+            slug
+            id
+          }
+          shop{
+            domain{
+                host
+                url
+            }
+          }
+          newEmail
+        }
+      }
+    }
+"""
+)
+
+
 ACCOUNT_DELETE_REQUESTED = (
     fragments.CUSTOMER_DETAILS
     + """