+++ CHANGELOG.md
@@ -49,7 +49,8 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add missing descriptions to Account module. - #13155 by @fowczarek
 - Add `ACCOUNT_CONFIRMATION_REQUESTED` async event - #13162 by @SzymJ
 - Add `ACCOUNT_DELETE_REQUESTED` async event - #13170 by @SzymJ
--
+- Add `ACCOUNT_CHANGE_EMAIL_REQUESTED` async event - #13233 by @SzymJ
+
 # 3.14.0
 
 ### Breaking changes